+++ CHANGELOG.rst
@@ -11,6 +11,10 @@ New features and enhancements
 * Add `additive_space` option to ``xs.train``. (:pull:`603`).
 * Add the option to `clip` before `additive_space`. (:pull:`604`).
 
+Bug fixes
+^^^^^^^^^
+* Better handling of attributes in ``xs.train``. (:pull:`608`, :issue:`607`)
+
 v0.12.3 (2025-05-26)
 --------------------
 Contributors to this version: Juliette Lavoie (:user:`juliettelavoie`).